# integrations/sre-alert-ingestion-service/openapi.yaml
openapi: 3.0.1
info:
  title: SRE Alert Ingestion Service API
  version: 1.0.0
  description: >
    Ingests normalized alerts from external monitoring/alerting tools and
    turns each into a platform incident. Every accepted alert is durably
    buffered before any delivery attempt is made, retried with backoff on
    failure, and escalated via an independent voice-call channel if delivery
    keeps failing — so this service is never a single point of failure on
    the platform's own availability. Deployed with no gateway/ingress auth
    layer in front of it, so it authenticates every inbound request itself
    via HTTP Basic Auth, one username/password pair per alert source.
security:
  - basicAuth: []
servers:
  - url: "{server}:{port}/"
    variables:
      server:
        default: http://localhost
      port:
        default: "8080"
paths:
...
  /alerts/adapters/aws:
    post:
      summary: Submit an AWS SNS-wrapped CloudWatch alarm notification.
      description: >
        Auto-confirms an SNS SubscriptionConfirmation/UnsubscribeConfirmation
        handshake message without creating an alert from either. A
        Notification message's own Message field carries the CloudWatch
        alarm as a JSON string, translated into this service's generic
        AlertRequest and enqueued via the exact same
        validation/buffering/worker/grouping/dedup/escalation path as
        POST /alerts.
      operationId: createAlertFromAWS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AWSSNSEnvelope'
      responses:
        "200":
          description: SNS control message (subscription/unsubscribe confirmation) handled; no alert created.
        "202":
          description: Alert accepted and durably buffered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertAcceptedResponse'
        "400":
          description: BadRequest — missing a required field, or malformed JSON.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "413":
          description: RequestEntityTooLarge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >
        One username/password pair per alert source, configured via the
        SRE_ALERT_AUTH_USERS environment variable (bcrypt-hashed at rest).
        See README.md's Authentication section.
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
...
    AlertAcceptedResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The buffered alert's internal id (not a platform incident id).
        alertNumber:
          type: string
          description: >
            This service's own human-readable alert number (e.g. "ALT0000123"),
            distinct from id — see internal/store.Store.Enqueue's doc comment.
...
    AWSSNSEnvelope:
      type: object
      description: >
        AWS SNS's own message envelope. Type distinguishes a real alarm
        notification from the two control messages SNS itself sends.
      properties:
        Type:
          type: string
          description: '"SubscriptionConfirmation", "UnsubscribeConfirmation", or "Notification".'
        MessageId:
          type: string
        SubscribeURL:
          type: string
          description: Present only on a SubscriptionConfirmation message; GETed to auto-confirm the subscription.
        Message:
          type: string
          description: >
            The CloudWatch alarm as a JSON string (not a nested object —
            SNS's own wire format). Its own AlarmDescription field carries
            a second, nested JSON string with service/severity/category/
            environment, since CloudWatch alarms have no native
            custom-fields mechanism.
...